refactor(codeclimate): reduce lines in file `src/main.ts`
Code Climate reported the following issue:

> File `main.ts` has 292 lines of code (exceeds 250 allowed). Consider refactoring.
