+++ .template/README.njk
@@ -104,7 +104,7 @@ Currently not supported.
 
 ### Default Defined Variables
 
-* `pkg` - [Source]({{ '/src/main.ts' | repoBrowseURL | linesSelectedURL( start=r/= *readPkgJson\(/,  end=r/Object\.assign\([a-zA-Z0-9]+, *\{ *pkg *\}\)|\.pkg *= *pkg *[;\n]/ ) }}) and [Source]({{ '/src/main.ts' | repoBrowseURL | linesSelectedURL( r/^function readPkgJson\((?:(?!\})[^\n]*\n)+/m ) }})
+* `pkg` - [Source]({{ '/src/main.ts' | repoBrowseURL | linesSelectedURL( start=r/= *readPkgJson\(/,  end=r/Object\.assign\([a-zA-Z0-9]+, *\{ *pkg *\}\)|\.pkg *= *pkg *[;\n]/ ) }}) and [Source]({{ '/src/utils/package-json.ts' | repoBrowseURL | linesSelectedURL( r/^export function readPkgJson\((?:(?!\})[^\n]*\n)+/m ) }})
 
     Object value of `package.json`
