+++ src/template-filters/repoBrowseURL.ts
@@ -2,8 +2,9 @@ import { dirname, relative as relativePath, resolve as resolvePath } from 'path'
 
 import type hostedGitInfo from 'hosted-git-info';
 
-import type { GetCommittishFn, GetHeadCommitSha1Fn, GetReleasedVersionsFn } from '../main';
+import type { GetHeadCommitSha1Fn, GetReleasedVersionsFn } from '../main';
 import { cachedPromise, errorMsgTag, isObject, validateString } from '../utils';
+import type { GetCommittishFn } from '../utils/hosted-git-info';
 import type { RepoData } from './linesSelectedURL';
 
 type RepoBrowseURLResult = RepoData & { gitRepoPath: string; toString: () => string };
